moves meeting date calculator
