+++ city_scrapers/spiders/cle_design_review.py
@@ -1,11 +1,11 @@
-import calendar
 import re
 import time
-from datetime import date, datetime, timedelta
+from datetime import datetime, timedelta
 
 from city_scrapers_core.constants import ADVISORY_COMMITTEE
 from city_scrapers_core.items import Meeting
 from city_scrapers_core.spiders import CityScrapersSpider
+from city_scrapers.utils import MeetingDateCalculator
 
 
 class CleDesignReviewSpider(CityScrapersSpider):
@@ -72,7 +72,7 @@ class CleDesignReviewSpider(CityScrapersSpider):
             location = self._parse_location(committee_meta)
             time_str = self._parse_time_str(committee_meta)
             email_contact = self._parse_email_contact(committee_meta)
-            weekday, chosen_weeks, is_downtown = self._parse_meeting_schedule_info(
+            weekday, chosen_ordinals, is_downtown = self._parse_meeting_schedule_info(
                 committee_meta
             )
             most_recent_start = datetime.today()
@@ -118,8 +118,8 @@ class CleDesignReviewSpider(CityScrapersSpider):
 
             calc_end = calc_start + timedelta(days=60)
 
-            upcoming_meetings = self._calculate_upcoming_meeting_days(
-                weekday, chosen_weeks, calc_start, calc_end
+            upcoming_meetings = MeetingDateCalculator.calculate_upcoming_meeting_days(
+                weekday, chosen_ordinals, calc_start, calc_end
             )
             if is_downtown:  # downtown meetings are a day before the one calculated
                 upcoming_meetings = [
@@ -231,13 +231,14 @@ class CleDesignReviewSpider(CityScrapersSpider):
 
         if is_downtown:
             weekday = 4
-            chosen_weeks = [0, 2]
+            chosen_ordinals = [0, 2]
         else:
             weekday_str = committee_meta.css("p.mb-1 strong::text").extract_first()
             weekday = self._parse_weekday(weekday_str)
             raw_weeks = re.findall(r"1st|2nd|3rd|4th", committee_str)
-            chosen_weeks = [self._parse_ordinal(ordinal) for ordinal in raw_weeks]
-        return weekday, chosen_weeks, is_downtown
+            # ordinals here just refer to the 1st, 2nd etc...
+            chosen_ordinals = [self._parse_ordinal(ordinal) for ordinal in raw_weeks]
+        return weekday, chosen_ordinals, is_downtown
 
     def _parse_weekday(self, weekday):
         """Parses weekday strings as their integer equivalent"""
@@ -250,91 +251,4 @@ class CleDesignReviewSpider(CityScrapersSpider):
         ordinal_lookup = {"1st": 0, "2nd": 1, "3rd": 2, "4th": 3}
         return ordinal_lookup[ordinal_str.lower()]
 
-    def _calculate_upcoming_meeting_days(
-        self, chosen_weekday, chosen_weeks, start, end
-    ):
-        """
-        This function is used to calculate meeting dates described as the 1st
-        and 3rd Tuesday of a month for any given time frame between start and
-        end dates.
-
-        Parameters:
-        chosen_weekday (int): the weekday that you're looking for. Monday is 0,
-            so in the examples above this would be 2
-        chosen_weeks (int[]): the particular days you're looking for - like 1st
-            and 3rd. These days should be passed though starting the count from 0,
-            i.e [0, 2] for first and third
-        start (date): the first day to begin calculating meetings from
-        end (date): the final day to be considered as a potential meeting date
-
-        Returns:
-        []date: an array of dates that match the given conditions
-        """
-        current_month = start.month
-        current_year = start.year
-
-        raw_dates = []
-        while not (current_month == end.month and current_year == end.year):
-            current_month_days = self._calculate_meeting_days_per_month(
-                chosen_weekday, chosen_weeks, current_year, current_month
-            )
-            raw_dates = raw_dates + [
-                date(current_year, current_month, day) for day in current_month_days
-            ]
-
-            # we can't easily use % arithmetic here since we're starting at 1, so
-            # it's a bit easier to read this way
-            current_month = current_month + 1 if current_month != 12 else 1
-            if current_month == 1:
-                current_year = current_year + 1
-
-        # add the days for the final month since they're missed by the loop
-        current_month_days = self._calculate_meeting_days_per_month(
-            chosen_weekday, chosen_weeks, current_year, current_month
-        )
-        raw_dates = raw_dates + [
-            date(current_year, current_month, day) for day in current_month_days
-        ]
-        # we now have all the relevant dates for the given months but we need to
-        # filter out days before and after start and end
-        return [
-            current_date
-            for current_date in raw_dates
-            if (start.date() <= current_date <= end.date())
-        ]
-
-    def _calculate_meeting_days_per_month(
-        self, chosen_weekday, chosen_weeks, year, month
-    ):
-        """
-        This function is used to calculate meeting dates described as the 1 and 3rd
-        Tuesday of a month.
-
-        Parameters:
-        chosen_weekday (int): the weekday that you're looking for. Monday is 0, so
-            in the examples above this would be 2
-        chosen_weeks (int[]): the particular days you're looking for - like 1st and
-            3rd. These days should be passed though starting the count from 0,
-            i.e [0, 2] for first and third
-        year (int): the year as an integer
-        month (int): the month as an integer
-
-        Returns:
-        []int: an array of the days of the month that matched the given conditions.
-        """
-
-        days_of_the_month = calendar.Calendar().itermonthdays2(year, month)
-        # we create a list of all days in the month that are the proper weekday -
-        # day is 0 if it is outside the month but present to make complete first or
-        # last weeks
-        potential_days = [
-            day
-            for day, weekday in days_of_the_month
-            if day != 0 and weekday == chosen_weekday
-        ]
-        # we then see if the resulting number is in the chosen_weeks array
-        chosen_days = [
-            day for i, day in enumerate(potential_days) if (i) in chosen_weeks
-        ]
-
-        return chosen_days
+
